move `isPostTypeHierarchical` in inspector controls component + reset parents on post type change
